+++ sdk-core/src/test/java/care/data4life/sdk/RecordServiceTest.java
@@ -26,7 +26,6 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
-import care.data4life.crypto.KeyType;
 import care.data4life.fhir.stu3.model.Attachment;
 import care.data4life.fhir.stu3.model.CarePlan;
 import care.data4life.fhir.stu3.model.DocumentReference;
@@ -41,10 +40,8 @@ import care.data4life.sdk.model.DeleteResult;
 import care.data4life.sdk.model.DownloadResult;
 import care.data4life.sdk.model.DownloadType;
 import care.data4life.sdk.model.Meta;
-import care.data4life.sdk.model.ModelVersion;
 import care.data4life.sdk.model.Record;
 import care.data4life.sdk.network.model.DecryptedRecord;
-import care.data4life.sdk.network.model.EncryptedRecord;
 import care.data4life.sdk.test.util.AttachmentBuilder;
 import care.data4life.sdk.util.Base64;
 import care.data4life.sdk.util.MimeType;
@@ -68,7 +65,6 @@ import static com.google.common.truth.Truth.assertThat;
 import static java.util.Arrays.asList;
 import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doReturn;
